[FAB2023-897] feat: interface 분리, node및edge style 분리, 노드 클릭 이벤트 기본 구조
